Fix some json variable types

+++ bundles/org.openhab.binding.teslapowerwall/src/main/java/org/openhab/binding/teslapowerwall/internal/api/MeterAggregates.java
@@ -53,7 +53,7 @@ public class MeterAggregates {
         public float frequency;
 
         @SerializedName("energy_exported")
-        public int energyExported;
+        public float energyExported;
 
         @SerializedName("energy_imported")
         public float energyImported;
